+++ crates/api/src/lib.rs
@@ -72,8 +72,6 @@ mod fields;
 mod filter;
 mod item_collection;
 mod items;
-#[cfg(feature = "python")]
-pub mod python;
 mod root;
 mod search;
 mod sort;