+++ crates/cli/src/lib.rs
@@ -257,10 +257,15 @@ struct ErrorLevel;
 
 impl Stacrs {
     /// Runs this command.
-    pub async fn run(self) -> Result<()> {
-        tracing_subscriber::fmt()
-            .with_max_level(self.log_level())
-            .init();
+    ///
+    /// If `init_tracing_subscriber` is `false`, it is expected that the caller
+    /// is setting up the appropriate logging (e.g. Python).
+    pub async fn run(self, init_tracing_subscriber: bool) -> Result<()> {
+        if init_tracing_subscriber {
+            tracing_subscriber::fmt()
+                .with_max_level(self.log_level())
+                .init();
+        }
         match self.command {
             Command::Translate {
                 ref infile,
@@ -700,7 +705,7 @@ mod tests {
         let stacrs = Stacrs::parse_from(["stacrs", "translate"]);
         assert_eq!(
             stacrs.input_format(Some("file.parquet")),
-            Format::Geoparquet(None)
+            Format::Geoparquet(Some(Compression::SNAPPY))
         );
 
         let stacrs = Stacrs::parse_from(["stacrs", "--input-format", "json", "translate"]);